// src/app/water/types.ts
export type Water = {
  latestDate: string
  name: string
  waterCurrentYear: number
  manpowerCurrentYear: number
  water_manpowerCurrentYear: number
  revenueCurrentYear: number
  billiRevenueCurrentYear: number
  waterCompareYear: number
  manpowerCompareYear: number
  water_manpowerCompareYear: number
  revenueCompareYear: number
  billiRevenueCompareYear: number
  waterWeight: number
  manpowerWeight: number
  water_manpowerWeight: number
  revenueWeight: number
  billiRevenueWeight: number
  waterGradient: number
  revenueGradient: number
  billiRevenueGradient: number
  compareBaseYear: number
  compareBaseGradient: number
  plants?: Water[]
}